Add files via upload
This report provides a comprehensive analysis of a fruit classification system which helps the local farmer to sort their fruit labels on the boxes. The system utilizes the Convolutional Neural Network (CNN), a powerful deep learning technique to classify different types of fruits. This dataset consist of various images which are categorized into 14 distinct fruit classes. This data is further inspect, preprocess and validate to enhance the model reliability.